Reverted the windows server to the original code--with changes to InputEvent instead of MouseEvent.
Testing concludes that the /entire/ windows lag problem has to do with disabling nagle on the client side.  It didn't have much effect on the linux client so I left nagle enabled since it's supposed to help... well regardless disabling it seems to make that horrible windows problem go away.